Create takediscountandnot.cpp
There are N items in a shop. You know that the price of the i-th item is Ai​
 . Chef wants to buy all the 
N items.

There is also a discount coupon that costs 

X rupees and reduces the cost of every item by 

Y rupees. If the price of an item was initially≤Y, it becomes free, i.e, costs 
0
0.

Determine whether Chef should buy the discount coupon or not. Chef will buy the discount coupon if and only if the total price he pays after buying the discount coupon is strictly less than the price he pays without buying the discount coupon.